requirements.txt
streamlit
selenium
transformers
torch
python-dotenv
pandas